refactor plugin to be more extensible
