+++ packages/next-intl/src/plugin/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './createNextIntlPlugin.tsx';
