+++ packages/next-intl/src/plugin/types.tsx
@@ -1,0 +1,3 @@
+export type PluginConfig = {
+  requestConfig?: string;
+};
